requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudscraper==1.2.60
commonmark==0.9.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.3
pandas==1.4.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.0.1
six==1.16.0
urllib3==1.26.9
yarl==1.7.2
